StringStore: Trim strings longer than 255 byte.
Can’t happen, you say? Well, apparently there are invalid UTF-8
sequences in the Wikipedia database for some comment strings, because
the last character got cut in the middle because of the 255 byte
database field limit.

What the XML export does is to convert that character to U+FFFD, which
is represented by 3 bytes of UTF-8. And voilà, your string is 257 bytes
long.

Sigh.